Added some base lniker code

+++ ErrorManager.h
@@ -69,7 +69,7 @@ namespace Astral
 		ss << std::setw(4) << lexeme.line + 1;
 		msg += ss.str() + "| " + ReplaceTabsWithSpaces(lexeme.lineData.c_str()) + '\n';
 
-		msg += std::string(StringLengthConvertTabToSpace(lexeme.lineData.c_str(), lexeme.positionInLine) + 6, ' ');
+		msg += std::string((size_t)StringLengthConvertTabToSpace(lexeme.lineData.c_str(), lexeme.positionInLine) + 6, ' ');
 		msg += std::string(lexeme.lexeme.size(), '^');
 
 		msg += "\n\n";
@@ -86,7 +86,7 @@ namespace Astral
 		ss << std::setw(4) << lexeme.line + 1;
 		msg += ss.str() + "| " + ReplaceTabsWithSpaces(lexeme.lineData.c_str()) + '\n';
 
-		msg += std::string(StringLengthConvertTabToSpace(lexeme.lineData.c_str(), lexeme.positionInLine) + 6, ' ');
+		msg += std::string((size_t)StringLengthConvertTabToSpace(lexeme.lineData.c_str(), lexeme.positionInLine) + 6, ' ');
 		msg += std::string(lexeme.lexeme.size(), '^');
 
 		msg += "\n\n";
